Fix object counts for classed objects
MT does not automatically insert into the `count()` method the proper terms for classed classes (e.g. Blog/Website) as it does with load, load_iter, et al.  This means that both `MT::Blog->count()` and `MT::Website->count()` will report the same number, namely the count of ALL of the records in the `mt_blog` table.

I'd call that a bug and here, we fix it, at least for count() on the objects.  `meta_count` is still affected.